+++ buildSrc/src/main/kotlin/ktx/Versions.kt
@@ -1,7 +1,7 @@
 package ktx
 
 const val gdxVersion = "1.12.1"
-const val kotlinCoroutinesVersion = "1.7.3"
+const val kotlinCoroutinesVersion = "1.8.1"
 
 const val artemisOdbVersion = "2.3.0"
 const val ashleyVersion = "1.7.4"
@@ -10,9 +10,9 @@ const val visUiVersion = "1.5.3"
 
 const val spekVersion = "1.1.5"
 const val kotlinTestVersion = "2.0.7"
-const val kotlinMockitoVersion = "4.1.0"
-const val assertjVersion = "3.11.1"
-const val junitVersion = "4.12"
-const val slf4jVersion = "1.7.26"
-const val wireMockVersion = "2.24.0"
+const val kotlinMockitoVersion = "5.4.0"
+const val assertjVersion = "3.26.3"
+const val junitVersion = "4.13.2"
+const val slf4jVersion = "1.7.36"
+const val wireMockVersion = "2.27.2"
 const val ktlintVersion = "0.50.0"